+++ packages/contracts/contracts/voting/addresslist/AddresslistVoting.sol
@@ -72,6 +72,11 @@ contract AddresslistVoting is Addresslist, MajorityVotingBase {
     }
 
     /// @inheritdoc MajorityVotingBase
+    function totalVotingPower(uint256 _blockNumber) public view override returns (uint256) {
+        return addresslistLengthAtBlock(_blockNumber);
+    }
+
+    /// @inheritdoc MajorityVotingBase
     function createProposal(
         bytes calldata _metadata,
         IDAO.Action[] calldata _actions,
@@ -108,8 +113,6 @@ contract AddresslistVoting is Addresslist, MajorityVotingBase {
         proposal_.parameters.votingMode = votingMode();
         proposal_.parameters.supportThreshold = supportThreshold();
         proposal_.parameters.minParticipation = minParticipation();
-
-        proposal_.tally.totalVotingPower = addresslistLengthAtBlock(snapshotBlock); // TODO https://aragonassociation.atlassian.net/browse/APP-1417
 
         for (uint256 i; i < _actions.length; ) {
             proposal_.actions.push(_actions[i]);