+++ packages/contracts/test/voting/token-voting.ts
@@ -382,9 +382,12 @@ describe('TokenVoting', function () {
         proposal.parameters.startDate.add(votingSettings.minDuration)
       ).to.equal(proposal.parameters.endDate);
 
-      expect(proposal.tally.totalVotingPower).to.equal(1);
       expect(proposal.tally.yes).to.equal(0);
       expect(proposal.tally.no).to.equal(0);
+
+      expect(
+        await voting.totalVotingPower(proposal.parameters.snapshotBlock)
+      ).to.equal(1);
 
       expect(
         await voting.canVote(1, signers[0].address, VoteOption.Yes)
@@ -439,10 +442,13 @@ describe('TokenVoting', function () {
       );
       expect(proposal.parameters.snapshotBlock).to.equal(block.number - 1);
 
-      expect(proposal.tally.totalVotingPower).to.equal(1);
       expect(proposal.tally.yes).to.equal(1);
       expect(proposal.tally.no).to.equal(0);
       expect(proposal.tally.abstain).to.equal(0);
+
+      expect(
+        await voting.totalVotingPower(proposal.parameters.snapshotBlock)
+      ).to.equal(1);
     });
 
     it('reverts creation when voting before the start date', async () => {