+++ packages/subgraph/tests/token-voting/token-voting.test.ts
@@ -27,7 +27,11 @@ import {
   TOTAL_VOTING_POWER
 } from '../constants';
 
-import {createDummyActions, createGetProposalCall} from '../utils';
+import {
+  createDummyActions,
+  createGetProposalCall,
+  createTotalVotingPowerCall
+} from '../utils';
 import {
   createNewVoteCastEvent,
   createNewProposalExecutedEvent,
@@ -65,9 +69,14 @@ test('Run TokenVoting (handleProposalCreated) mappings with mock event', () => {
     '0', // abstain
     '0', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
+  );
+
+  createTotalVotingPowerCall(
+    CONTRACT_ADDRESS,
+    SNAPSHOT_BLOCK,
+    TOTAL_VOTING_POWER
   );
 
   // create event
@@ -177,9 +186,14 @@ test('Run TokenVoting (handleVoteCast) mappings with mock event', () => {
     '0', // abstain
     '1', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
+  );
+
+  createTotalVotingPowerCall(
+    CONTRACT_ADDRESS,
+    SNAPSHOT_BLOCK,
+    TOTAL_VOTING_POWER
   );
 
   // create event
@@ -251,7 +265,6 @@ test('Run TokenVoting (handleVoteCast) mappings with mock event', () => {
     '0', // abstain
     '2', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
   );
@@ -306,7 +319,6 @@ test('Run TokenVoting (handleVoteCast) mappings with mock event and vote option 
     '0', // abstain
     '0', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
   );
@@ -358,7 +370,6 @@ test('Run TokenVoting (handleProposalExecuted) mappings with mock event', () => 
     '0', // abstain
     '1', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
   );